refactor(backend): move tmdb services into one module
