+++ apps/backend/src/media/resolver.rs
@@ -46,9 +46,11 @@ use crate::{
         VideoGameSpecifics,
     },
     providers::{
-        audible::AudibleService, igdb::IgdbService, listennotes::ListennotesService,
-        movies_tmdb::MovieTmdbService, openlibrary::OpenlibraryService,
-        shows_tmdb::ShowTmdbService,
+        audible::AudibleService,
+        igdb::IgdbService,
+        listennotes::ListennotesService,
+        openlibrary::OpenlibraryService,
+        tmdb::{MovieTmdbService, ShowTmdbService},
     },
     traits::MediaProvider,
     utils::{user_auth_token_from_ctx, user_id_from_ctx, MemoryDb, NamedObject},
