+++ apps/backend/src/providers/mod.rs
@@ -1,6 +1,5 @@
 pub mod audible;
 pub mod igdb;
 pub mod listennotes;
-pub mod movies_tmdb;
 pub mod openlibrary;
-pub mod shows_tmdb;
+pub mod tmdb;
